refactor: split to module for maintainability feat: eq options
