+++ nameless/command/music/player_manager.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import logging
-from typing import TYPE_CHECKING, cast
+from typing import TYPE_CHECKING, Literal, cast, overload
 
 import discord
 from discord.ext import commands
@@ -20,15 +20,63 @@ class PlayerManager:
     def __init__(self, bot: Nameless):
         self.bot = bot
 
-    async def validate_voice_permissions(self, ctx: commands.Context[Nameless]) -> None:
+    async def validate_voice_permissions(
+        self,
+        ctx: commands.Context[Nameless],
+        *,
+        bypass_checks: bool = False,
+    ) -> None:
         if not isinstance(ctx.author, discord.Member):
             raise commands.CommandError("This command can only be used in a server!")
+
+        if bypass_checks:
+            return
 
         if not ctx.author.voice or not ctx.author.voice.channel:
             raise NotInVoiceError()
 
+    @overload
     async def get_player(
-        self, ctx: commands.Context[Nameless], *, connect_if_needed: bool = True
+        self,
+        ctx: commands.Context[Nameless],
+        *,
+        connect_if_needed: bool = False,
+        raise_on_unconnected: Literal[True],
+    ) -> CustomPlayer: ...
+
+    @overload
+    async def get_player(
+        self,
+        ctx: commands.Context[Nameless],
+        *,
+        connect_if_needed: Literal[True],
+        raise_on_unconnected: Literal[True],
+    ) -> CustomPlayer: ...
+
+    @overload
+    async def get_player(
+        self,
+        ctx: commands.Context[Nameless],
+        *,
+        connect_if_needed: Literal[True],
+        raise_on_unconnected: Literal[False] = False,
+    ) -> CustomPlayer | None: ...
+
+    @overload
+    async def get_player(
+        self,
+        ctx: commands.Context[Nameless],
+        *,
+        connect_if_needed: Literal[False] = False,
+        raise_on_unconnected: Literal[False] = False,
+    ) -> CustomPlayer | None: ...
+
+    async def get_player(
+        self,
+        ctx: commands.Context[Nameless],
+        *,
+        connect_if_needed: bool = False,
+        raise_on_unconnected: bool = False,
     ) -> CustomPlayer | None:
         if not ctx.guild:
             raise commands.CommandError("This command can only be used in a server!")
@@ -38,13 +86,16 @@ class PlayerManager:
                 raise ValueError("Voice client is not a CustomPlayer instance!")
             return ctx.guild.voice_client
 
+        if raise_on_unconnected:
+            raise ConnectionFailedError("No active voice connection found.")
+
         if connect_if_needed:
             return await self.connect_to_voice(ctx)
 
         return None
 
     async def get_or_create_player(self, ctx: commands.Context[Nameless]) -> CustomPlayer:
-        player = await self.get_player(ctx, connect_if_needed=True)
+        player = await self.get_player(ctx, connect_if_needed=True, raise_on_unconnected=False)
         if player is None:
             raise ConnectionFailedError("Could not establish voice connection")
         return player
@@ -53,8 +104,9 @@ class PlayerManager:
         self,
         ctx: commands.Context[Nameless],
         channel: discord.VoiceChannel | discord.StageChannel | None = None,
+        bypass_checks: bool = False,
     ) -> CustomPlayer | None:
-        await self.validate_voice_permissions(ctx)
+        await self.validate_voice_permissions(ctx, bypass_checks=bypass_checks)
 
         if channel is None and isinstance(ctx.author, discord.Member):
             if ctx.author.voice and ctx.author.voice.channel:
